+++ app/controllers/concerns/authenticable_user.rb
@@ -4,6 +4,8 @@ module AuthenticableUser
   def current_user
     return unless token && payload
 
+    unauthenticated_request_error unless payload && refresh_token
+
     User.find_by(id: payload["sub"])
   end
 
@@ -15,5 +17,19 @@ module AuthenticableUser
     @payload ||= JWT.decode(token, ENV.fetch("AUTH_SECRET_TOKEN"), true, algorithm: "HS256").first
   rescue JWT::DecodeError
     nil
+  end
+
+  def jti
+    payload["jti"]
+  end
+
+  def refresh_token
+    RefreshToken.active.find_by(jti: jti).present?
+  end
+
+  def unauthenticated_request_error
+    GraphQL::ExecutionError.new(
+      extensions: { title: "Invalid credentials", detail: "", status: 401, code: :unauthorized, meta: {} }
+    )
   end
 end
