+++ app/interactors/create_access_token.rb
@@ -3,17 +3,11 @@ class CreateAccessToken
 
   ACCESS_TOKEN_TTL = 1.hour
 
-  delegate :user, :client_uid, to: :context
-
-  before do
-    context.client_uid ||= client_uid
-  end
+  delegate :user, to: :context
 
   def call
     context.fail!(error_data: error_data) unless user
-
     context.access_token = access_token
-    context.client_uid = client_uid
     context.jti = jti
   end
 
@@ -27,7 +21,6 @@ class CreateAccessToken
     {
       sub: user.id,
       exp: ACCESS_TOKEN_TTL.from_now.to_i,
-      client_uid: client_uid,
       jti: jti,
       type: "access"
     }
@@ -35,10 +28,6 @@ class CreateAccessToken
 
   def error_data
     { message: "Invalid credentials", status: 401, code: :unauthorized }
-  end
-
-  def client_uid
-    "#{user.id}-#{Time.current.to_i}"
   end
 
   def jti
